requirements.txt
flask==2.0.1
flask-login==0.5.0
boto3==1.26.137
python-dotenv==0.19.0
psutil==5.9.5
requests==2.31.0
discord.py==2.3.2
gunicorn==21.2.0
werkzeug==2.0.3 